platform: Update info command with region
Updates the `platform info` command to use a region on the API call to
the SDK.
